Eliminated compilation warning.
git-svn-id: http://crystal.svn.sourceforge.net/svnroot/crystal/CS/trunk@21118 8cc4aa7f-3514-0410-904f-f2cc9021211c

+++ plugins/culling/dynavis/exvis.cpp
@@ -41,7 +41,7 @@ csExactCuller::csExactCuller (int w, int h)
   z_buffer = new float[w*h];
   for (int i = 0 ; i < w*h ; i++)
   {
-    scr_buffer[i] = ~0;
+    scr_buffer[i] = (uint32)~0;
     z_buffer[i] = 1000000000000.0f;
   }
